refactor(portal): move shared Color types to scenes/common/colors
The Color/ColorName/colors type+const are used by v2 (Header, Profile
ColorSelector, color-atom) and v3 (UserPopup), so they move from
Profile/types.ts to scenes/common/colors. Pure relocation plus import
repoints; calculateColorFromString stays in @/lib and the jotai colorAtom
stays in Portal/layout.
